# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/shifts": {
      "post": {
        "operationId": "ShiftController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShiftDto"
              },
              "examples": {
                "default": {
                  "summary": "Morning shift",
                  "value": {
                    "name": "Morning",
                    "startTime": "08:00",
                    "endTime": "17:00",
                    "gracePeriodMinutes": 15
                  }
                },
                "overnight": {
                  "summary": "Night shift (crosses midnight)",
                  "value": {
                    "name": "Night",
                    "startTime": "22:00",
                    "endTime": "06:00",
                    "gracePeriodMinutes": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Shifts"
        ]
      },
      "get": {
        "operationId": "ShiftController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Shifts"
        ]
      }
    },
    "/shifts/{id}": {
      "get": {
        "operationId": "ShiftController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Shifts"
        ]
      },
      "patch": {
        "operationId": "ShiftController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShiftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Shifts"
        ]
      },
      "delete": {
        "operationId": "ShiftController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Shifts"
        ]
      }
    },
...
    "/employees/{employeeId}/salary-components": {
      "post": {
        "operationId": "SalaryComponentController_create",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSalaryComponentDto"
              },
              "examples": {
                "allowance": {
                  "summary": "Fixed housing allowance",
                  "value": {
                    "type": "ALLOWANCE",
                    "name": "Housing",
                    "amount": 1000,
                    "isPercentage": false
                  }
                },
                "percentage": {
                  "summary": "Percentage-based allowance",
                  "value": {
                    "type": "ALLOWANCE",
                    "name": "Transport",
                    "amount": 10,
                    "isPercentage": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Salary Components"
        ]
      },
      "get": {
        "operationId": "SalaryComponentController_findAll",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Salary Components"
        ]
      }
    },
    "/salary-components/{id}": {
      "patch": {
        "operationId": "SalaryComponentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSalaryComponentDto"
              },
              "examples": {
                "default": {
                  "summary": "Update amount",
                  "value": {
                    "amount": 1500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Salary Components"
        ]
      },
      "delete": {
        "operationId": "SalaryComponentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Salary Components"
        ]
      }
    },
...
  },
  "info": {
    "title": "HR System API",
    "description": "Multi-tenant Payroll & HR Management API (Phases 1–6). Import into APIDog via live URL http://localhost:3004/docs-json or backend/postman_collection.json. Swagger UI: http://localhost:3004/docs",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3004/api",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "schemas": {
...
      "CreateShiftDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Morning",
            "maxLength": 120
          },
          "startTime": {
            "type": "string",
            "example": "08:00",
            "description": "24-hour HH:mm"
          },
          "endTime": {
            "type": "string",
            "example": "17:00",
            "description": "24-hour HH:mm. May be <= startTime for overnight shifts."
          },
          "gracePeriodMinutes": {
            "type": "number",
            "example": 15,
            "minimum": 0,
            "maximum": 240
          }
        },
        "required": [
          "name",
          "startTime",
          "endTime"
        ]
      },
      "UpdateShiftDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Morning",
            "maxLength": 120
          },
          "startTime": {
            "type": "string",
            "example": "08:00"
          },
          "endTime": {
            "type": "string",
            "example": "17:00"
          },
          "gracePeriodMinutes": {
            "type": "number",
            "example": 15,
            "minimum": 0,
            "maximum": 240
          }
        }
      },
...
      "CreateSalaryComponentDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "ALLOWANCE",
              "DEDUCTION"
            ],
            "example": "ALLOWANCE"
          },
          "name": {
            "type": "string",
            "example": "Housing",
            "maxLength": 120
          },
          "amount": {
            "type": "number",
            "example": 1000,
            "minimum": 0,
            "description": "Fixed amount, or a 0–100 percentage when isPercentage is true."
          },
          "isPercentage": {
            "type": "boolean",
            "example": false,
            "description": "When true, amount is a percentage of the basic salary."
          }
        },
        "required": [
          "type",
          "name",
          "amount"
        ]
      },
      "UpdateSalaryComponentDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "ALLOWANCE",
              "DEDUCTION"
            ]
          },
          "name": {
            "type": "string",
            "example": "Transport",
            "maxLength": 120
          },
          "amount": {
            "type": "number",
            "example": 500,
            "minimum": 0
          },
          "isPercentage": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
    }
  }
}